Template installing backend working

+++ app/libraries/Helper.php
@@ -77,17 +77,20 @@ class Helper {
 
     public static function deleteFolderAndContents($dir)
     {
-        $it = new RecursiveDirectoryIterator($dir);
-        $files = new RecursiveIteratorIterator($it,
-            RecursiveIteratorIterator::CHILD_FIRST);
-        foreach($files as $file) {
-            if ($file->getFilename() === '.' || $file->getFilename() === '..') {
-                continue;
-            }
-            if ($file->isDir()){
-                rmdir($file->getRealPath());
-            } else {
-                unlink($file->getRealPath());
+        if(file_exists($dir))
+        {
+            $it = new RecursiveDirectoryIterator($dir);
+            $files = new RecursiveIteratorIterator($it,
+                RecursiveIteratorIterator::CHILD_FIRST);
+            foreach($files as $file) {
+                if ($file->getFilename() === '.' || $file->getFilename() === '..') {
+                    continue;
+                }
+                if ($file->isDir()){
+                    rmdir($file->getRealPath());
+                } else {
+                    unlink($file->getRealPath());
+                }
             }
         }
     }
